backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
boto3>=1.34.129
requests-oauthlib>=2.0.0
cryptography>=42.0.8
python-dotenv>=1.0.1
pymongo==4.5.0
pydantic>=2.6.4
pydantic[email]>=2.6.4
email-validator>=2.2.0
pyjwt>=2.10.1
passlib[bcrypt]>=1.7.4
tzdata>=2024.2
motor==3.3.1
pytest>=8.0.0
black>=24.1.1
isort>=5.13.2
flake8>=7.0.0
mypy>=1.8.0
python-jose[cryptography]>=3.3.0
requests>=2.31.0
pandas>=2.2.0
numpy>=1.26.0
python-multipart>=0.0.9
jq>=1.6.0
typer>=0.9.0
bcrypt>=4.1.2
emergentintegrations>=0.1.0
sentry-sdk[fastapi]>=1.40.0
apscheduler>=3.10.0
paypalrestsdk>=1.2.0
stripe>=7.0.0
aiofiles>=23.2.1
google-generativeai>=0.3.0
langchain>=0.1.0
openai>=1.0.0
redis>=5.0.0
celery>=5.3.0
flower>=2.0.0
prometheus-client>=0.19.0
slowapi>=0.1.9
